Added a file exists check to copy_thumbnails
Former-commit-id: 4dc9214952150b246ad26b3c77a6e4c379d488d1

+++ Litao/copy_thumbnails.py
@@ -28,6 +28,9 @@ def copy_images_to_dir(animal, channel, resolution):
         dst_file = os.path.join(file_location_manager.prep, channel_dir,
                                 resolution, str(section.section_number).zfill(3) + '.tif')
 
+        if os.path.exists(dst_file):
+            continue
+
         os.makedirs(os.path.dirname(dst_file), exist_ok=True)
         copyfile(src_file, dst_file)
